+++ pkg/ref/random_test.go
@@ -101,3 +101,53 @@ func TestRandom(t *testing.T) {
 
 	assert.NoError(t, r.Validate())
 }
+
+func TestRandomWithoutBooks(t *testing.T) {
+	t.Parallel()
+
+	r, err := ref.Random(
+		ref.ExcludeReferences(
+			"Matthew", "Mark", "Luke", "John", "Acts",
+			"Romans", "1 Corinthians", "2 Corinthians",
+			"Galatians", "Ephesians", "Philippians",
+			"Colossians", "1 Thessalonians", "2 Thessalonians",
+			"1 Timothy", "2 Timothy", "Titus", "Philemon",
+			"Hebrews", "James", "1 Peter", "2 Peter", "1 John", "2 John", "3 John", "Jude",
+			"Revelation",
+		),
+	)
+
+	otBooks := []string{
+		"Genesis", "Exodus", "Leviticus", "Numbers", "Deuteronomy",
+		"Joshua", "Judges", "Ruth", "1 Samuel", "2 Samuel", "1 Kings", "2 Kings",
+		"1 Chronicles", "2 Chronicles", "Ezra", "Nehemiah", "Esther", "Job", "Psalms",
+		"Proverbs", "Ecclesiastes", "Song of Songs", "Isaiah", "Jeremiah",
+		"Lamentations", "Ezekiel", "Daniel", "Hosea", "Joel", "Amos", "Obadiah",
+		"Jonah", "Micah", "Nahum", "Habakkuk", "Zephaniah", "Haggai", "Zechariah",
+		"Malachi",
+	}
+
+	assert.NoError(t, err)
+	assert.NotNil(t, r)
+
+	assert.NoError(t, r.Validate())
+	assert.Contains(t, otBooks, r.Book.Name)
+}
+
+func TestRandomWithoutBooksFromCategory(t *testing.T) {
+	t.Parallel()
+
+	r, err := ref.Random(
+		ref.ExcludeReferences(
+			"Matthew", "Mark",
+		),
+		ref.FromCategory("Gospels"),
+	)
+
+	assert.NoError(t, err)
+	assert.NotNil(t, r)
+
+	assert.NoError(t, r.Validate())
+	assert.NotEqual(t, "Matthew", r.Book.Name)
+	assert.NotEqual(t, "Mark", r.Book.Name)
+}