+++ Changes.md
@@ -1,8 +1,15 @@
 ## WIP  TBD
 
+ * :computer: The `today ost index` command has been added to pull down scripture indexes from openscripture.today.
+ * :computer: Added `--exclude` and `--exclude-index` options to `today random`.
  * ESV text from VerseText and others won't include the references now.
  * Updating the Verse and Photo file formats used by the `ost.Client` for the upcoming version of openscripture.today. Adding a `ost.Metadata` structure that records version, original verseion, and whether the verse and photo have been pruned.
  * A new `VersesIndex` has been added to the `ost.Client` for fetching indexes listing all the verses that have been postd for certain periods (all time, year, month).
+ * Added a `Clone` method to `ref.Canon` and `ref.Book` to create deep clones of these objects.
+ * Added the `ref.ExcludeReferences` option for use with the `ref.Random` functions.
+ * Added a `Subtract` method to `ref.Resolved` to return the difference between two resolved references.
+ * Added a `Filtered` method to `ref.Canon` that returns a new canon that has segments references removed.
+ * :boom: Breaking Change :boom:: Now requires Go 1.22.
  * :boom: Breaking Change :boom:: Significant changes have been made to the photos API. These changes include the following:
    - The `photo.Meta` and `photo.Info` structures have been removed/merged/refactored into `photo.Descriptor`.
    - A new interface named `photo.Image` has been added and each `photo.Descriptor` should have at least one, but may have many associated `photo.Image` objects.
@@ -26,6 +33,8 @@
  * New utility function `unsplash.IDFromURL()` added which will give you the photo ID from an Unsplash photo URL.
  * :hammer: Fix: Multiple total chapter references will be resolved to chapter ranges. For example, if you parse "Ps. 12-13", the resolver will correctly return "Psalms 12-13" instead of "Psalms 12:1-13:6" as it would have before.
  * :hammer: Fix: Psalms singular handling has been special cased so that references to a single verse can show something like "Psalm 12" or "Psalm 12:1-3" rather than "Psalms 12" or "Psalms 12:1-3" as it would have before.
+ * :hammer: Fix: Photo descriptors were not being correctly written to JSON files prior to this version. They were missing the "creator" key and embedding the name of the creator in the structure above. This has been corrected.
+ * :hammer: Fix: Verse JSON files were not written correctly either prior to this version. They were missing the "version" key and embedding the name of the version in the parent structure. This has been fixed.
 
 ## 0.5.1  2024-02-20